Fix gcc warning

+++ ll-opengl/source/examples/point-based-example/main.cpp
@@ -218,7 +218,7 @@ int main()
         gl::glEnable(gl::GL_BLEND);
         gl::glBlendFunc(gl::GL_SRC_ALPHA, gl::GL_ONE_MINUS_SRC_ALPHA);
 
-		for (int i = 0; i < vertexClouds.size(); i++) {
+		for (size_t i = 0; i < vertexClouds.size(); i++) {
 			renderer.render(vertexClouds[i]);
 		}
